+++ src/main/java/jumper/Constants.java
@@ -4,7 +4,6 @@
 
 package jumper;
 
-import java.util.List;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
@@ -104,8 +103,6 @@ public class Constants {
   public static final String TOKEN_CLAIM_ACCESS_TOKEN_ENVIRONMENT = "env";
   public static final String TOKEN_CLAIM_ACCESS_TOKEN_PUBLISHER_ID = "publisherId";
   public static final String TOKEN_CLAIM_ACCESS_TOKEN_SUBSCRIBER_ID = "subscriberId";
-
-  public static final List<String> SPACE_ZONES = List.of("space", "canis", "aries");
 
   public static final String BASIC_AUTH_PROVIDER_KEY = "default";
   public static final String OAUTH_PROVIDER_KEY = "default";