+++ src/test/java/jumper/config/CucumberContextConfiguration.java
@@ -7,9 +7,11 @@ package jumper.config;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
 
 @io.cucumber.spring.CucumberContextConfiguration
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureWebTestClient(timeout = "PT10S") // PT65S - PT = Period time, S = seconds
+@ContextConfiguration(classes = {CustomHttpClientConfiguration.class})
 public class CucumberContextConfiguration {}